// apps/api/src/modules/iam/auth.schema.ts
import { z } from "zod";

export const loginSchema = z.object({
  email: z.string().email(),
  password: z.string().min(6),
});

export type LoginInput = z.infer<typeof loginSchema>;

export const registerSchema = z.object({
  email: z.string().email(),
  password: z.string().min(6),
  fullName: z.string().min(2),
  role: z.enum(["ADVISOR", "ADMIN"]).default("ADVISOR"),
});

export type RegisterInput = z.infer<typeof registerSchema>;